digital-organism/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
numpy==1.24.3
torch==2.1.0
transformers==4.35.0
datasets==2.14.6
pydantic==2.5.0
tinker-ai
accelerate==0.25.0
python-dotenv==1.0.0
